Simplify FixSLinksPatch(es)
